自动更新部分，除p2_fundlist,沪深300指外，其它一天更新一次。修改理财师更新部分为增量更新

+++ application/controllers/systemsetup/AutoUpdate.php
@@ -13,35 +13,72 @@ class AutoUpdate extends MY_Controller {
 	
 	public function index() {
 		$this->fund_interface->autoUpdateJZInfo();		//更新p2_fundlist信息
-		$tableNames = array('hsindexvalue'=>'fund_hsindexvalue','fundmanager'=>'fund_manager','fundmanagerinfo'=>'fund_managerinfo',
-							'funddistribution'=>'fund_distribution','fundposition'=>'fund_position');
-		foreach ($tableNames as $key => $val){			//更新基金其它信息
-			$returnData = $this->fund_interface->autoUpdateJZInfo($val);
-			if (!empty($returnData['data'])){
-// 				$this->db->truncate($key);
-				$this->db->where('1=1')->delete($key);
-				$this->db->insert_batch($key,$returnData['data']);
+//------------------------更新沪深300指数据-----------------------
+		$returnData = $this->fund_interface->autoUpdateJZInfo('fund_hsindexvalue',array('TradingDay>='=>date('Y-m-d',time()-2592000)));		
+		$this->load->model("Model_db");
+		$this->Model_db->incremenUpdate('p2_hsindexvalue',$returnData['data'],'TradingDay');
+		$this->updateHsindexCurve();
+//----------------更新基金经理人，基金资产分布，基金重仓股信息--------------------
+		$updateRecode = $this->db->where(array('dealitem' => 'fundTables'))->get('dealitems')->row_array();
+		if (empty($updateRecode)){
+			$this->db->set(array('dealitem' => 'fundTables','updatetime' => 0 ))->insert('dealitems');
+			$updatetime = 0;
+		}else{
+			$updatetime = $updateRecode['updatetime'];
+		}
+		if (time()-$updatetime > 86400){
+			$tableNames = array('fundmanager'=>'fund_manager','fundmanagerinfo'=>'fund_managerinfo',
+					'funddistribution'=>'fund_distribution','fundposition'=>'fund_position');
+			$flag = true;
+			foreach ($tableNames as $key => $val){			//更新基金其它信息
+				$returnData = $this->fund_interface->autoUpdateJZInfo($val);
+				if (!empty($returnData['data'])){
+					$this->db->where('1=1')->delete($key);
+					$this->db->insert_batch($key,$returnData['data']);
+					$dbErr = $this->db->error();
+					$flag = $flag && !$dbErr['code'];
+				}
 			}
-			if ($key == 'hsindexvalue'){
-				$this->updateHsindexCurve();
+			if ($flag){
+				$this->db->set(array('updatetime' => time()))->where(array('dealitem' => 'fundTables'))->update('dealitems');
 			}
 		}
-		$this->updatePlannerInfo();						//更新理财师信息
+//--------------更新理财师信息-------------------------------------		
+		$updateRecode = $this->db->where(array('dealitem' => 'plannerInfo'))->get('dealitems')->row_array();
+		if (empty($updateRecode)){
+			$this->db->set(array('dealitem' => 'plannerInfo','updatetime' => 0 ))->insert('dealitems');
+			$updatetime = 0;
+		}else{
+			$updatetime = $updateRecode['updatetime'];
+		}
+		if (time()-$updatetime > 86400){
+			if ($this->updatePlannerInfo()){						//更新理财师信息
+				$this->db->set(array('updatetime' => time()))->where(array('dealitem' => 'plannerInfo'))->update('dealitems');
+			}
+		}
 	}
 
 	function updatePlannerInfo(){
 		$this->load->config('jz_dict');
 		$plannerInfo = json_decode(comm_curl($this->config->item('XNPlannerUrl'),array()),true)[0];
+// var_dump($plannerInfo);
 		if ('0000' == $plannerInfo['code']){
 			$newData = array();
+			$flag = true;
 			foreach ($plannerInfo['data'] as &$val){
-				$newData[] = array('FName'=>$val['name'],'EmployeeID'=>$val['workNum'],'status'=>1,'area'=>$val['area'],'city'=>$val['city']); 
+				$status = 1;
+				if(isset($val['available']) && !$val['available']){
+					$status = 0;
+				}
+				$newData[] = array('FName'=>$val['name'],'EmployeeID'=>$val['workNum'],'status'=>$status,'area'=>$val['area'],'city'=>$val['city']); 
 			}
 			if (!empty($newData)){
-				$this->load->model("Model_db");
-				$this->db->set(array('status'=>0))->update('p2_planner');
-				$this->Model_db->incremenUpdate('p2_planner',$newData,'EmployeeID');
+//				$this->db->set(array('status'=>0))->update('p2_planner');
+				$flag = $flag && $this->Model_db->incremenUpdate('p2_planner',$newData,'EmployeeID');
 			}
+			return $flag;
+		}else{
+			return false;
 		}
 	}
